+++ pkg/controller/constants/constants.go
@@ -1,6 +1,7 @@
 package constants
 
 import (
+	"strings"
 	"time"
 
 	"github.com/infinispan/infinispan-operator/pkg/controller/utils/common"
@@ -10,6 +11,9 @@ import (
 var (
 	// DefaultImageName is used if a specific image name is not provided
 	DefaultImageName = common.GetEnvWithDefault("DEFAULT_IMAGE", "infinispan/server:latest")
+
+	// JGroupsDiagnosticsFlag is used to enable traces for JGroups
+	JGroupsDiagnosticsFlag = strings.ToUpper(common.GetEnvWithDefault("JGROUPS_DIAGNOSTICS", "FALSE"))
 
 	// DefaultMemorySize string with default size for memory
 	DefaultMemorySize = resource.MustParse("512Mi")
@@ -34,7 +38,10 @@ const (
 	// DefaultDeveloperUser users to access the cluster rest API
 	DefaultDeveloperUser = "developer"
 	// DefaultCacheName default cache name for the CacheService
-	DefaultCacheName = "default"
+	DefaultCacheName   = "default"
+	InfinispanPort     = 11222
+	InfinispanPingPort = 8888
+	CrossSitePort      = 7900
 	// DefaultCacheManagerName default cache manager name used for cross site
 	DefaultCacheManagerName                 = "default"
 	CacheServiceFixedMemoryXmxMb            = 200
@@ -42,12 +49,25 @@ const (
 	CacheServiceMaxRamMb                    = CacheServiceFixedMemoryXmxMb + CacheServiceJvmNativeMb
 	CacheServiceAdditionalJavaOptions       = "-Dsun.zip.disableMemoryMapping=true -XX:+UseSerialGC -XX:MinHeapFreeRatio=5 -XX:MaxHeapFreeRatio=10"
 	CacheServiceJvmNativePercentageOverhead = 1
-	InfinispanFinalizer                     = "finalizer.infinispan.org"
 
 	ServerHTTPBasePath         = "rest/v2"
 	ServerHTTPClusterStop      = ServerHTTPBasePath + "/cluster?action=stop"
 	ServerHTTPHealthPath       = ServerHTTPBasePath + "/cache-managers/default/health"
 	ServerHTTPHealthStatusPath = ServerHTTPHealthPath + "/status"
+
+	DefaultCacheTemplate = `<infinispan>
+		<cache-container>
+			<distributed-cache name="%v" mode="SYNC" owners="%d" statistics="true">
+				<memory>
+					<off-heap size="%d" eviction="MEMORY" strategy="REMOVE"/>
+				</memory>
+				<partition-handling when-split="ALLOW_READ_WRITES" merge-policy="REMOVE_ALL" />
+			</distributed-cache>
+		</cache-container>
+	</infinispan>`
+
+	// Using for local run and test only
+	ClusterUpKubeConfig = "../../openshift.local.clusterup/kube-apiserver/admin.kubeconfig"
 )
 
 const (
