+++ pkg/controller/infinispan/util/cluster.go
@@ -58,7 +58,7 @@ func (c Cluster) GracefulShutdown(user, pass, podName, namespace, protocol strin
 	if err != nil {
 		return err
 	}
-	httpURL := fmt.Sprintf("%s://%v:11222/%s", protocol, podIP, consts.ServerHTTPClusterStop)
+	httpURL := fmt.Sprintf("%s://%v:%d/%s", protocol, podIP, consts.InfinispanPort, consts.ServerHTTPClusterStop)
 	commands := []string{"curl", "-X", "GET", "--insecure", "-u", fmt.Sprintf("%v:%v", user, pass), httpURL}
 
 	logger := log.WithValues("Request.Namespace", namespace, "Pod.Name", podName)
@@ -89,7 +89,7 @@ func (c Cluster) GetClusterMembers(user, pass, podName, namespace, protocol stri
 		return nil, err
 	}
 
-	httpURL := fmt.Sprintf("%s://%v:11222/%s", protocol, podIP, consts.ServerHTTPHealthPath)
+	httpURL := fmt.Sprintf("%s://%v:%d/%s", protocol, podIP, consts.InfinispanPort, consts.ServerHTTPHealthPath)
 	commands := []string{"curl", "--insecure", "-u", fmt.Sprintf("%v:%v", user, pass), httpURL}
 
 	logger := log.WithValues("Request.Namespace", namespace, "Pod.Name", podName)
@@ -119,11 +119,11 @@ func (c Cluster) ExistsCache(user, pass, cacheName, podName, namespace, protocol
 		return false, err
 	}
 
-	httpURL := fmt.Sprintf("%s://%s:11222/rest/v2/caches/%s", protocol, podIP, cacheName)
+	httpURL := fmt.Sprintf("%s://%s:%d/rest/v2/caches/%s?action=config", protocol, podIP, consts.InfinispanPort, cacheName)
 	commands := []string{"curl",
 		"--insecure",
 		"--http1.1",
-		"-so", "/dev/null", "-w", "%{http_code}", // ignore output and get http status code
+		"-o", "/dev/null", "-w", "%{http_code}", // ignore output and get http status code
 		"-u", fmt.Sprintf("%v:%v", user, pass),
 		"--head",
 		httpURL,
@@ -156,7 +156,7 @@ func (c Cluster) CacheNames(user, pass, podName, namespace, protocol string) ([]
 	if err != nil {
 		return nil, err
 	}
-	httpURL := fmt.Sprintf("%s://%s:11222/rest/v2/caches", protocol, podIP)
+	httpURL := fmt.Sprintf("%s://%s:%d/rest/v2/caches", protocol, podIP, consts.InfinispanPort)
 	commands := []string{"curl", "--insecure", "-u", fmt.Sprintf("%v:%v", user, pass), httpURL}
 	logger := log.WithValues("Request.Namespace", namespace, "Pod.Name", podName)
 	logger.Info("get caches list", "url", httpURL)
@@ -178,7 +178,7 @@ func (c Cluster) CreateCacheWithTemplate(user, pass, cacheName, cacheXML, podNam
 		return err
 	}
 
-	httpURL := fmt.Sprintf("%s://%s:11222/rest/v2/caches/%s", protocol, podIP, cacheName)
+	httpURL := fmt.Sprintf("%s://%s:%d/rest/v2/caches/%s", protocol, podIP, consts.InfinispanPort, cacheName)
 	commands := []string{"curl",
 		"--insecure",
 		"--http1.1",
@@ -220,7 +220,7 @@ func (c Cluster) CreateCacheWithTemplateName(user, pass, cacheName, templateName
 		return err
 	}
 
-	httpURL := fmt.Sprintf("%s://%s:11222/rest/v2/caches/%s?template=%s", protocol, podIP, cacheName, templateName)
+	httpURL := fmt.Sprintf("%s://%s:%d/rest/v2/caches/%s?template=%s", protocol, podIP, consts.InfinispanPort, cacheName, templateName)
 	commands := []string{"curl",
 		"--insecure",
 		"--http1.1",
@@ -307,7 +307,7 @@ func (c Cluster) GetMetrics(user, pass, podName, namespace, protocol, postfix st
 	if err != nil {
 		return nil, err
 	}
-	httpURL := fmt.Sprintf("%s://%s:11222/metrics"+postfix, protocol, podIP)
+	httpURL := fmt.Sprintf("%s://%s:%d/metrics/%s", protocol, podIP, consts.InfinispanPort, postfix)
 	commands := []string{"curl", "--insecure", "--http1.1",
 		"-u", fmt.Sprintf("%v:%v", user, pass),
 		"-H", "Accept: application/json", httpURL}
@@ -327,7 +327,7 @@ func ClusterStatusHandler(scheme corev1.URIScheme) corev1.Handler {
 		HTTPGet: &corev1.HTTPGetAction{
 			Scheme: scheme,
 			Path:   consts.ServerHTTPHealthStatusPath,
-			Port:   intstr.FromInt(11222)},
+			Port:   intstr.FromInt(consts.InfinispanPort)},
 	}
 }
 
@@ -337,7 +337,7 @@ func (c Cluster) GetCacheManagerInfo(user, pass, cacheManagerName, podName, name
 	if err != nil {
 		return nil, err
 	}
-	httpURL := fmt.Sprintf("%s://%s:11222/rest/v2/cache-managers/%s", protocol, podIP, cacheManagerName)
+	httpURL := fmt.Sprintf("%s://%s:%d/rest/v2/cache-managers/%s", protocol, podIP, consts.InfinispanPort, cacheManagerName)
 	commands := []string{"curl", "--insecure", "-u", fmt.Sprintf("%v:%v", user, pass), httpURL}
 	logger := log.WithValues("Request.Namespace", namespace, "Pod.Name", podName)
 	logger.Info("get cache-manager info", "url", httpURL)
