+++ test/e2e/main_test.go
@@ -17,7 +17,8 @@ import (
 	"github.com/infinispan/infinispan-operator/pkg/controller/infinispan/util"
 	comutil "github.com/infinispan/infinispan-operator/pkg/controller/utils/common"
 	tconst "github.com/infinispan/infinispan-operator/test/e2e/constants"
-	testutil "github.com/infinispan/infinispan-operator/test/e2e/util"
+	k8s "github.com/infinispan/infinispan-operator/test/e2e/k8s"
+	testutils "github.com/infinispan/infinispan-operator/test/e2e/utils"
 	"gopkg.in/yaml.v2"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -27,7 +28,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 )
 
-var kubernetes = testutil.NewTestKubernetes()
+var kubernetes = k8s.NewTestKubernetes()
 var cluster = util.NewCluster(kubernetes.Kubernetes)
 
 var DefaultSpec = ispnv1.Infinispan{
@@ -65,7 +66,7 @@ var MinimalSpec = ispnv1.Infinispan{
 
 func TestMain(m *testing.M) {
 	namespace := strings.ToLower(tconst.Namespace)
-	if "true" == tconst.RunLocalOperator {
+	if "TRUE" == tconst.RunLocalOperator {
 		kubernetes.DeleteNamespace(namespace)
 		kubernetes.DeleteCRD("infinispans.infinispan.org")
 		kubernetes.DeleteCRD("caches.infinispan.org")
@@ -252,7 +253,7 @@ func TestCacheService(t *testing.T) {
 
 	user := cconsts.DefaultDeveloperUser
 	password, err := cluster.Kubernetes.GetPassword(user, spec.GetSecretName(), tconst.Namespace)
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 
 	ispn := ispnv1.Infinispan{}
 	kubernetes.Kubernetes.Client.Get(context.TODO(), types.NamespacedName{Namespace: spec.Namespace, Name: spec.Name}, &ispn)
@@ -287,7 +288,7 @@ func TestPermanentCache(t *testing.T) {
 	// Define function for the generic stop/start test procedure
 	var createPermanentCache = func(ispn *ispnv1.Infinispan) {
 		pass, err := cluster.Kubernetes.GetPassword(usr, ispn.GetSecretName(), tconst.Namespace)
-		testutil.ExpectNoError(err)
+		testutils.ExpectNoError(err)
 		routeName := fmt.Sprintf("%s-external", name)
 		protocol := getSchemaForRest(ispn)
 
@@ -301,7 +302,7 @@ func TestPermanentCache(t *testing.T) {
 
 	var usePermanentCache = func(ispn *ispnv1.Infinispan) {
 		pass, err := cluster.Kubernetes.GetPassword(usr, ispn.GetSecretName(), tconst.Namespace)
-		testutil.ExpectNoError(err)
+		testutils.ExpectNoError(err)
 		routeName := fmt.Sprintf("%s-external", name)
 		protocol := getSchemaForRest(ispn)
 
@@ -338,7 +339,7 @@ func TestCheckDataSurviveToShutdown(t *testing.T) {
 	// Define function for the generic stop/start test procedure
 	var createCacheWithFileStore = func(ispn *ispnv1.Infinispan) {
 		pass, err := cluster.Kubernetes.GetPassword(usr, ispn.GetSecretName(), tconst.Namespace)
-		testutil.ExpectNoError(err)
+		testutils.ExpectNoError(err)
 		routeName := fmt.Sprintf("%s-external", name)
 		protocol := getSchemaForRest(ispn)
 
@@ -354,7 +355,7 @@ func TestCheckDataSurviveToShutdown(t *testing.T) {
 
 	var useCacheWithFileStore = func(ispn *ispnv1.Infinispan) {
 		pass, err := cluster.Kubernetes.GetPassword(usr, ispn.GetSecretName(), tconst.Namespace)
-		testutil.ExpectNoError(err)
+		testutils.ExpectNoError(err)
 		routeName := fmt.Sprintf("%s-external", name)
 		schema := getSchemaForRest(ispn)
 		tr := &http.Transport{
@@ -405,7 +406,7 @@ func waitForPodsOrFail(spec *ispnv1.Infinispan, num int) {
 	podName := pods[0].Name
 
 	pass, err := cluster.Kubernetes.GetPassword(cconsts.DefaultOperatorUser, spec.GetSecretName(), tconst.Namespace)
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 
 	expectedClusterSize := num
 	// Check that the cluster size is num querying the first pod
@@ -427,7 +428,7 @@ func waitForPodsOrFail(spec *ispnv1.Infinispan, num int) {
 		err = fmt.Errorf("timed out waiting for the condition, last error: %v", lastErr)
 	}
 
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 }
 
 func TestExternalService(t *testing.T) {
@@ -461,7 +462,7 @@ func TestExternalService(t *testing.T) {
 	kubernetes.WaitForPods("app=infinispan-pod", 1, tconst.SinglePodTimeout, tconst.Namespace)
 
 	pass, err := cluster.Kubernetes.GetPassword(usr, spec.GetSecretName(), tconst.Namespace)
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 
 	routeName := fmt.Sprintf("%s-external", name)
 	ispn := ispnv1.Infinispan{}
@@ -515,7 +516,7 @@ func TestExternalServiceWithAuth(t *testing.T) {
 	newpass := "connectornewpass"
 	identities := util.CreateIdentitiesFor(usr, pass)
 	identitiesYaml, err := yaml.Marshal(identities)
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 
 	// Create secret with application credentials
 	secret := corev1.Secret{
@@ -565,7 +566,7 @@ func TestExternalServiceWithAuth(t *testing.T) {
 	// Update the auth credentials.
 	identities = util.CreateIdentitiesFor(usr, newpass)
 	identitiesYaml, err = yaml.Marshal(identities)
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 
 	// Create secret with application credentials
 	secret1 := corev1.Secret{
@@ -705,11 +706,11 @@ func httpEmpty(httpURL string, method string, usr string, pass string, flags str
 	if flags != "" {
 		req.Header.Set("flags", flags)
 	}
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 
 	req.SetBasicAuth(usr, pass)
 	resp, err := client.Do(req)
-	testutil.ExpectNoError(err)
+	testutils.ExpectNoError(err)
 
 	if resp.StatusCode != http.StatusOK {
 		panic(httpError{resp.StatusCode})
