+++ test/e2e/utils/asserts.go
@@ -1,4 +1,4 @@
-package util
+package utils
 
 import "k8s.io/apimachinery/pkg/api/errors"
